requirements.txt
# Core web framework
fastapi==0.109.1
uvicorn[standard]==0.25.0

# Database and ORM
sqlalchemy==2.0.23
async-sqlalchemy==0.0.1
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment variables
python-dotenv==1.0.0

# HTTP client for external APIs
httpx==0.25.2

# Background tasks
celery==5.3.4
redis==5.0.1

# Audio processing
pydub==0.25.1
librosa==0.10.1

# AWS SDK
boto3==1.34.15

# AI and NLP
openai==1.10.0
anthropic==0.17.0
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.25
tiktoken==0.5.2

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-slugify==8.0.1
cryptography==41.0.8

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Image processing
Pillow==10.1.0